BottomNavigationDataProvider was moved to test.data
