+++ jdi-light-material-ui-tests/src/main/java/io/github/com/custom/elements/CollapsedBreadcrumbs.java
@@ -1,0 +1,14 @@
+package io.github.com.custom.elements;
+
+import com.epam.jdi.light.elements.pageobjects.annotations.locators.UI;
+import com.epam.jdi.light.material.elements.navigation.Breadcrumbs;
+import com.epam.jdi.light.ui.html.elements.common.Button;
+
+public class CollapsedBreadcrumbs extends Breadcrumbs {
+    @UI(".MuiButtonBase-root")
+    protected Button expandButton;
+
+    public void expand() {
+        expandButton.click();
+    }
+}
