+++ jdi-light-material-ui-tests/src/test/java/io/github/epam/material/tests/navigation/BottomNavigationTests.java
@@ -1,8 +1,8 @@
 package io.github.epam.material.tests.navigation;
 
 import io.github.epam.TestsInit;
+import io.github.epam.test.data.BottomNavigationDataProvider;
 import org.testng.annotations.BeforeMethod;
-import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 
 import static io.github.com.StaticSite.bottomNavigationPage;
@@ -18,16 +18,7 @@ public class BottomNavigationTests extends TestsInit {
         bottomNavigationPage.checkOpened();
     }
 
-    @DataProvider(name = "bottomNavigationButtons")
-    public static Object[][] bottomNavigationButtons() {
-        return new Object[][] {
-                {1, "Recents", "isRecents(0)"},
-                {2, "Favorites", "isFavorites(1)"},
-                {3, "Nearby", "isNearby(2)"},
-        };
-    }
-
-    @Test(dataProvider = "bottomNavigationButtons")
+    @Test(dataProvider = "bottomNavigationButtons", dataProviderClass = BottomNavigationDataProvider.class)
     public void defaultBottomNavigationTest(int index, String buttonText, String positionText) {
         bottomNavigationItems.select(index);
         bottomNavigationItems.has().selected(index);
