+++ jdi-light-material-ui-tests/src/test/java/io/github/epam/material/tests/navigation/SimpleBreadcrumbsTests.java
@@ -57,9 +57,7 @@ public class SimpleBreadcrumbsTests extends TestsInit {
     @Test
     public void collapsedBreadcrumbsTest() {
         collapsedBreadcrumbs.has().values("Home", "Belts");
-        // TODO: Add expand method
-
-        collapsedBreadcrumbs.find(".MuiButtonBase-root").click();
+        collapsedBreadcrumbs.expand();
         collapsedBreadcrumbs.has().values("Home", "Catalog", "Accessories", "New Collection", "Belts");
     }
