--- jdi-light-material-ui/src/main/java/com/epam/jdi/light/material/elements/navigation/CollapsedBreadcrumbs.java
@@ -1,7 +1,0 @@
-package com.epam.jdi.light.material.elements.navigation;
-
-import com.epam.jdi.light.elements.complex.dropdown.Dropdown;
-
-public class CollapsedBreadcrumbs extends Dropdown {
-
-}
